agnoTeams/pyproject.toml
[project]
name = "agnoteams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.8",
    "ddgs>=9.9.3",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
]
